mqtt-Use FPP as default hostname if not set playlist pause -support floating point number of seconds. Fixes #452 Don't display piface gpio inputs on BBB

+++ src/playlist/PlaylistEntryPause.h
@@ -46,7 +46,7 @@ class PlaylistEntryPause : public PlaylistEntryBase {
 	Json::Value GetConfig(void);
 
   private:
-	int                  m_duration;
+	float                m_duration;
 	long long            m_startTime;
 	long long            m_endTime;
 	long long            m_finishTime;
